@@ -55,6 +55,7 @@ class JavaSecurityCrawler < CommonSecurity
 
 
   def self.update sv, yml, affected
+    sv.source      = 'victims-cve-db'
     sv.description = yml['description']
     sv.summary     = yml['title']
     sv.summary     = sv.name_id if sv.summary.to_s.empty?
@@ -63,7 +64,11 @@ class JavaSecurityCrawler < CommonSecurity
     sv.affected_versions_string = affected['version'].to_a.join(" && ")
     sv.patched_versions_string  = affected['fixedin'].to_a.join(" && ")
     yml["references"].to_a.each do |reference|
-      key = reference.gsub(".", ":")
+      key = reference.gsub(".", "::")
+      match = reference.match(/(CVE-.*)\z/i)
+      if match
+        key = match[0].gsub(/(\?.*)\z/, "").gsub(".", "_")
+      end
       if sv.links && !sv.links.values.include?(reference)
         sv.links[key] = reference
       end
