@@ -40,6 +40,7 @@ class RubySecurityCrawler < CommonSecurity
     name_id  = filepath.split("/").last.gsub(".yaml", "").gsub(".yml", "")
 
     sv              = fetch_sv Product::A_LANGUAGE_RUBY, prod_key, name_id
+    sv.source       = 'ruby-advisory-db'
     sv.summary      = yml['title']
     sv.description  = yml['description']
     sv.framework    = yml['framework']
@@ -48,7 +49,9 @@ class RubySecurityCrawler < CommonSecurity
     sv.cvss_v2      = yml['cvss_v2']
     sv.osvdb        = yml['osvdb']
     sv.publish_date = yml['date']
-    sv.links['URL'] = yml['url']
+    if !sv.links.values.include?( yml['url'] )
+      sv.links['URL'] = yml['url']
+    end
     sv.unaffected_versions_string = yml['unaffected_versions'].to_a.join('||')
     sv.patched_versions_string    = yml['patched_versions'].to_a.join('||')
 
@@ -64,12 +67,13 @@ class RubySecurityCrawler < CommonSecurity
     product = sv.product
     return nil if product.nil?
 
-    unaffected_1 = VersionService.from_or_ranges product.versions, sv.unaffected_versions_string
-    unaffected_2 = VersionService.from_or_ranges product.versions, sv.patched_versions_string
+    unaffected = VersionService.from_or_ranges product.versions, sv.unaffected_versions_string
+    patched    = VersionService.from_or_ranges product.versions, sv.patched_versions_string
 
     affected_versions = []
     product.versions.each do |version|
-      next if unaffected_1.to_a.include?(version) || unaffected_2.to_a.include?(version)
+      next if unaffected.to_a.map(&:to_s).include?(version.to_s)
+      next if patched.to_a.map(&:to_s).include?(version.to_s)
       affected_versions << version
     end
 
